Fit a natural effect model

This function is similar to R function neModel from the medflex package.

ds.neModel(
  formula = NULL,
  family = NULL,
  expData = NULL,
  se = "robust",
  nBoot = 1000,
  newobj = NULL,
  datasources = NULL
)

Arguments

formula
a formula object providing a symbolic description of the natural effect model.

family
a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported.

expData
the expanded dataset (of class "expData").

se
character string indicating the type of standard errors to be calculated. By default robust se are calculated. NOTE: In version 1.0.0 the "bootstrap" method does not work properly

nBoot
number of bootstrap replicates.

newobj
a character string that provides the name for the output object that is stored on the data servers. Default neModel.out.

datasources
a list of DSConnection-class objects obtained after login. If the datasources argument is not specified the default set of connections will be used: see datashield.connections_default.

Value

a summary table of the object of class 'neModel'.

Details

The function ds.neModel is used to fit a natural effect model on the expanded dataset. For more details see the help file of the neModel function in the medflex package.

Author

Demetris Avraam, for DataSHIELD Development Team